+++ packages/core/navigation-next/src/components/GlobalItem/primitives.js
@@ -109,12 +109,13 @@ class GlobalNavigationItemPrimitive extends Component<*> {
   };
 
   render() {
-    const { isActive, isSelected, tooltip } = this.props;
+    const { isSelected, tooltip } = this.props;
     return (
       <Tooltip
         delay={0}
-        content={isSelected || isActive ? undefined : tooltip}
+        content={isSelected ? undefined : tooltip}
         position="right"
+        hideTooltipOnClick
         hideTooltipOnMouseDown
       >
         <div className={css({ display: 'inline-block' })}>