+++ packages/core/navigation-next/src/components/GlobalNav/index.js
@@ -7,6 +7,7 @@
 
 import React, { Component } from 'react';
 
+import GlobalItem from '../GlobalItem';
 import { withGlobalTheme } from '../../theme';
 import GlobalNavigation from './GlobalNavigation';
 import type { ConnectedGlobalNavigationProps } from './types';
@@ -16,6 +17,9 @@ const GlobalNavigationWithTheme = withGlobalTheme(GlobalNavigation);
 export default class ConnectedGlobalNavigation extends Component<
   ConnectedGlobalNavigationProps,
 > {
+  static defaultProps = {
+    itemComponent: GlobalItem,
+  };
   render() {
     return <GlobalNavigationWithTheme {...this.props} />;
   }