+++ packages/css-packs/reduced-ui-pack/src/__tests__/unit/index.test.js
@@ -1,0 +1,11 @@
+//@flow
+import reducedUiPackStyles from '../..';
+
+describe('styleSheet', () => {
+  it('should not have [object Object]', () => {
+    expect(reducedUiPackStyles).not.toContain('[object Object]');
+  });
+  it('should not have [Function', () => {
+    expect(reducedUiPackStyles).not.toContain('[Function');
+  });
+});